+++ modules/backup/r-blob-policy.tf
@@ -1,0 +1,8 @@
+resource "azurerm_data_protection_backup_policy_blob_storage" "blob_policy" {
+  count = var.backup_storage_blob_enabled ? 1 : 0
+
+  name     = local.blob_policy_name
+  vault_id = azurerm_data_protection_backup_vault.vault[0].id
+
+  retention_duration = "P${var.storage_blob_backup_policy_retention_in_days}D"
+}