+++ modules/backup/r-logs.tf
@@ -1,8 +1,10 @@
 module "diagnostics" {
   source  = "claranet/diagnostic-settings/azurerm"
-  version = "6.2.0"
+  version = "~> 6.2.0"
 
-  resource_id = azurerm_recovery_services_vault.vault.id
+  count = var.backup_vm_enabled || var.backup_file_share_enabled ? 1 : 0
+
+  resource_id = azurerm_recovery_services_vault.vault[0].id
 
   logs_destinations_ids = var.logs_destinations_ids
   log_categories        = var.logs_categories
@@ -13,4 +15,9 @@ module "diagnostics" {
   custom_name    = var.custom_diagnostic_settings_name
   name_prefix    = var.name_prefix
   name_suffix    = var.name_suffix
+}
+
+moved {
+  from = module.diagnostics
+  to   = module.diagnostics[0]
 }