+++ modules/logs/variables-naming.tf
@@ -42,12 +42,6 @@ variable "logs_storage_account_custom_name" {
   default     = ""
 }
 
-variable "logs_storage_account_appservices_container_name" {
-  description = "Name of the container in which App Services logs are stored"
-  type        = string
-  default     = "app-services"
-}
-
 variable "logs_storage_account_archived_logs_fileshare_name" {
   description = "Name of the file share in which externalized logs are stored"
   type        = string