+++ variables.tf
@@ -36,9 +36,3 @@ variable "location_short" {
   description = "Short string for Azure location."
   type        = string
 }
-
-variable "iaas_features_enabled" {
-  description = "Whether IaaS (Virtual Machine related) modules are created."
-  type        = bool
-  default     = true
-}