+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Unreleased
+
+Breaking
+  * AZ-926: Cleaning and refactoring
+
+Added
+  * AZ-926: Add Backup Vault resources
+
+Removed
+  * AZ-926: Remove Log Storage SAS token exposition
+  * AZ-926: Remove legacy IIS log ingestion configuration
+
 # v7.0.0 - 2023-01-13
 
 Added